Archive Redirects
This commit ports over the existing .htaccess file for the mith.umd.edu
website. In addition to continuing to redirect to the places where
things have moved to it will send all 404s over to
https://archive.mith.umd.edu/mith-2020/.

As of 2020 the mith.umd.edu had a very flat URL namespace, where 1283
paths lived at the root. Rather than explicitly listing all these out in
the .htaccess Apache will temporarily redirect requests for anything
that is missing over to the archive. Once the new static site is stable
we may want to make these permanent. We may also want to just list out
the thousands of paths here if it proves problematic to send all 404s to
the archive.