gatsby-remark-source-name
Since we have multiple sources for markdown now (airtable and the
filesystem) we need to make sure that our GraphQL queries are specific
to the node type we are interested in. The gatsby-remark-source-name
plugin lets you access the sourceName of the markdown node. This commit
adjusts the queries for posts to limit to the sourceName news, which is
defined in node-config.js