Flattened /news/ namespace
Prior to this commit the Airtable posts table had slugs without a year
in them e.g. `/news/{slug}`. However the markdown files in the github
repository were placed in a subdirectory named after the year of
publication: `/news/{year}/{slug}`. This allowed for duplicate
slugs to exist on the file system, which would map to Airtable with
unpredictable results. To solve this all the news markdown files are now in `src/news`
without the year subdirectory.

Moving the files actually uncovered a bug in the GraphQL query in
post-index.js which was attempting to connect the markdown files and
Airtable posts. Since the link wasn't always being made this resulted in
many empty pages.

This query has been updated to page through the Airtable post data,
sorting by date, and then connecting up each post with the excerpt from
the appropriate markdown files. Previously it was keying off of the
order of the markdown files, and attempting to connect those with
Airtable while limiting results, which wasn't returning 25 items per
page.