Website link and font-awesome imports
I noticed that homepage URLs for people in Airtable lacked the https:// prefix so
clicking on them in people pages resulted in a 404 error. I also noticed
that the websites lacked a font-awesome icon. I went ahead and added one
fa-globe, and also modified the font-awesome library to only include the
things we are using, which might cut down on the size of the app bundle.