Upgrade dependencies
I happened to notice that ERROR #11329 was printing out during the build after
I had added the google analytics plugin. I didn't notice it at first because
the build was completing successfully. I found some conversation about
how this error goes away after upgrading gatsby. I did a `npm update` which updated
a few versions in the package.json and package-lock.json.

With the new versions the build then alerted me to an invalid option
`background` for the gatsby-plugin-sharp module in the gatsby-config.js. It
does look like background was never doing anything based on the documentation, so
I removed it and things appear to be working.

You will probably want to do a `npm i` when you pull this change.