Pedantic markdown parsing
It seems that Gatsby defaults to parsing Markdown using Remark in
pedantic mode. Turning it off explicitly in gatsby-config.js should
help situations where we have links with underscores.

https://github.com/remarkjs/remark/issues/470

It seemed like the general consensus on GitHub was that pedantic should
be off by default, but that people were worried about breaking backwards
compatibility. Hopefully turning pedantic off will fix situations where
we had links containing underscores, but we should be on the lookout for
other changes to our HTML.